+++ packages/minimongo/minimongo_tests_client.js
@@ -137,47 +137,47 @@ Tinytest.add('minimongo - basics', test => {
   test.equal(c.find('abc').count(), 0);
   test.equal(c.find(undefined).count(), 0);
   test.equal(c.find().count(), 3);
-  test.equal(c.find(1, {skip: 1}).count(), 1);
-  test.equal(c.find(1, {skip: 1}).count(true), 0);
-  test.equal(c.find({_id: 1}, {skip: 1}).count(), 1);
-  test.equal(c.find({_id: 1}, {skip: 1}).count(true), 0);
+  test.equal(c.find(1, {skip: 1}).count(false), 1);
+  test.equal(c.find(1, {skip: 1}).count(), 0);
+  test.equal(c.find({_id: 1}, {skip: 1}).count(false), 1);
+  test.equal(c.find({_id: 1}, {skip: 1}).count(), 0);
   test.equal(c.find({_id: undefined}).count(), 0);
   test.equal(c.find({_id: false}).count(), 0);
   test.equal(c.find({_id: null}).count(), 0);
   test.equal(c.find({_id: ''}).count(), 0);
   test.equal(c.find({_id: 0}).count(), 0);
-  test.equal(c.find({}, {skip: 1}).count(), 3);
-  test.equal(c.find({}, {skip: 1}).count(true), 2);
-  test.equal(c.find({}, {skip: 2}).count(true), 1);
-  test.equal(c.find({}, {limit: 2}).count(), 3);
-  test.equal(c.find({}, {limit: 2}).count(true), 2);
-  test.equal(c.find({}, {limit: 1}).count(true), 1);
-  test.equal(c.find({}, {skip: 1, limit: 1}).count(), 3);
-  test.equal(c.find({}, {skip: 1, limit: 1}).count(true), 1);
-  test.equal(c.find({tags: 'fruit'}, {skip: 1}).count(), 2);
-  test.equal(c.find({tags: 'fruit'}, {skip: 1}).count(true), 1);
-  test.equal(c.find({tags: 'fruit'}, {limit: 1}).count(), 2);
-  test.equal(c.find({tags: 'fruit'}, {limit: 1}).count(true), 1);
-  test.equal(c.find({tags: 'fruit'}, {skip: 1, limit: 1}).count(), 2);
-  test.equal(c.find({tags: 'fruit'}, {skip: 1, limit: 1}).count(true), 1);
-  test.equal(c.find(1, {sort: ['_id', 'desc'], skip: 1}).count(), 1);
-  test.equal(c.find(1, {sort: ['_id', 'desc'], skip: 1}).count(true), 0);
-  test.equal(c.find({_id: 1}, {sort: ['_id', 'desc'], skip: 1}).count(), 1);
-  test.equal(c.find({_id: 1}, {sort: ['_id', 'desc'], skip: 1}).count(true), 0);
-  test.equal(c.find({}, {sort: ['_id', 'desc'], skip: 1}).count(), 3);
-  test.equal(c.find({}, {sort: ['_id', 'desc'], skip: 1}).count(true), 2);
-  test.equal(c.find({}, {sort: ['_id', 'desc'], skip: 2}).count(true), 1);
-  test.equal(c.find({}, {sort: ['_id', 'desc'], limit: 2}).count(), 3);
-  test.equal(c.find({}, {sort: ['_id', 'desc'], limit: 2}).count(true), 2);
-  test.equal(c.find({}, {sort: ['_id', 'desc'], limit: 1}).count(true), 1);
-  test.equal(c.find({}, {sort: ['_id', 'desc'], skip: 1, limit: 1}).count(), 3);
-  test.equal(c.find({}, {sort: ['_id', 'desc'], skip: 1, limit: 1}).count(true), 1);
-  test.equal(c.find({tags: 'fruit'}, {sort: ['_id', 'desc'], skip: 1}).count(), 2);
-  test.equal(c.find({tags: 'fruit'}, {sort: ['_id', 'desc'], skip: 1}).count(true), 1);
-  test.equal(c.find({tags: 'fruit'}, {sort: ['_id', 'desc'], limit: 1}).count(), 2);
-  test.equal(c.find({tags: 'fruit'}, {sort: ['_id', 'desc'], limit: 1}).count(true), 1);
-  test.equal(c.find({tags: 'fruit'}, {sort: ['_id', 'desc'], skip: 1, limit: 1}).count(), 2);
-  test.equal(c.find({tags: 'fruit'}, {sort: ['_id', 'desc'], skip: 1, limit: 1}).count(true), 1);
+  test.equal(c.find({}, {skip: 1}).count(false), 3);
+  test.equal(c.find({}, {skip: 1}).count(), 2);
+  test.equal(c.find({}, {skip: 2}).count(), 1);
+  test.equal(c.find({}, {limit: 2}).count(false), 3);
+  test.equal(c.find({}, {limit: 2}).count(), 2);
+  test.equal(c.find({}, {limit: 1}).count(), 1);
+  test.equal(c.find({}, {skip: 1, limit: 1}).count(false), 3);
+  test.equal(c.find({}, {skip: 1, limit: 1}).count(), 1);
+  test.equal(c.find({tags: 'fruit'}, {skip: 1}).count(false), 2);
+  test.equal(c.find({tags: 'fruit'}, {skip: 1}).count(), 1);
+  test.equal(c.find({tags: 'fruit'}, {limit: 1}).count(false), 2);
+  test.equal(c.find({tags: 'fruit'}, {limit: 1}).count(), 1);
+  test.equal(c.find({tags: 'fruit'}, {skip: 1, limit: 1}).count(false), 2);
+  test.equal(c.find({tags: 'fruit'}, {skip: 1, limit: 1}).count(), 1);
+  test.equal(c.find(1, {sort: ['_id', 'desc'], skip: 1}).count(false), 1);
+  test.equal(c.find(1, {sort: ['_id', 'desc'], skip: 1}).count(), 0);
+  test.equal(c.find({_id: 1}, {sort: ['_id', 'desc'], skip: 1}).count(false), 1);
+  test.equal(c.find({_id: 1}, {sort: ['_id', 'desc'], skip: 1}).count(), 0);
+  test.equal(c.find({}, {sort: ['_id', 'desc'], skip: 1}).count(false), 3);
+  test.equal(c.find({}, {sort: ['_id', 'desc'], skip: 1}).count(), 2);
+  test.equal(c.find({}, {sort: ['_id', 'desc'], skip: 2}).count(), 1);
+  test.equal(c.find({}, {sort: ['_id', 'desc'], limit: 2}).count(false), 3);
+  test.equal(c.find({}, {sort: ['_id', 'desc'], limit: 2}).count(), 2);
+  test.equal(c.find({}, {sort: ['_id', 'desc'], limit: 1}).count(), 1);
+  test.equal(c.find({}, {sort: ['_id', 'desc'], skip: 1, limit: 1}).count(false), 3);
+  test.equal(c.find({}, {sort: ['_id', 'desc'], skip: 1, limit: 1}).count(), 1);
+  test.equal(c.find({tags: 'fruit'}, {sort: ['_id', 'desc'], skip: 1}).count(false), 2);
+  test.equal(c.find({tags: 'fruit'}, {sort: ['_id', 'desc'], skip: 1}).count(), 1);
+  test.equal(c.find({tags: 'fruit'}, {sort: ['_id', 'desc'], limit: 1}).count(false), 2);
+  test.equal(c.find({tags: 'fruit'}, {sort: ['_id', 'desc'], limit: 1}).count(), 1);
+  test.equal(c.find({tags: 'fruit'}, {sort: ['_id', 'desc'], skip: 1, limit: 1}).count(false), 2);
+  test.equal(c.find({tags: 'fruit'}, {sort: ['_id', 'desc'], skip: 1, limit: 1}).count(), 1);
 
   // Regression test for #455.
   c.insert({foo: {bar: 'baz'}});
@@ -3466,8 +3466,8 @@ Tinytest.add('minimongo - count on cursor with limit', test => {
 
   const c = Tracker.autorun(c => {
     const cursor = coll.find({_id: {$exists: true}}, {sort: {_id: 1}, limit: 3});
-    count = cursor.count(true);
-    unlimitedCount = cursor.count();
+    count = cursor.count();
+    unlimitedCount = cursor.count(false);
   });
 
   test.equal(count, 3);
@@ -3774,8 +3774,8 @@ Tinytest.add('minimongo - reactive skip/limit count while updating', test => {
   let unlimitedCount = -1;
 
   const c = Tracker.autorun(() => {
-    count = X.find({}, {skip: 1, limit: 1}).count(true);
-    unlimitedCount = X.find({}, {skip: 1, limit: 1}).count();
+    count = X.find({}, {skip: 1, limit: 1}).count();
+    unlimitedCount = X.find({}, {skip: 1, limit: 1}).count(false);
   });
 
   test.equal(count, 0);